+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## [master]
+- Repl: +/. in p2 score snippets supported
+- Repl: + in p3 score snippet supported
+- Score: `csound-score--align-cols` improved to decrease col width as well
+
+
 ## [0.1.2]
 - Font-lock: Single line comments within score repaired
 - Font-lock: p-field variables in light color mode are darker